Removed HttpInvocationHelper again and cleaned up code
