+++ connectors/http-json/src/test/java/io/camunda/connector/http/LocalConnectorRuntime.java
@@ -18,9 +18,7 @@ package io.camunda.connector.http;
 
 import io.camunda.connector.runtime.ConnectorRuntimeApplication;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-//@SpringBootApplication
 public class LocalConnectorRuntime {
 
   public static void main(String[] args) {
